+++ mayan/apps/navigation/classes.py
@@ -24,7 +24,7 @@ from mayan.apps.common.literals import (
     TEXT_SORT_ORDER_PARAMETER, TEXT_SORT_ORDER_VARIABLE_NAME
 )
 from mayan.apps.common.settings import setting_home_view
-from mayan.apps.common.utils import resolve_attribute
+from mayan.apps.common.utils import get_related_field, resolve_attribute
 from mayan.apps.permissions import Permission
 
 from .html_widgets import SourceColumnLinkWidget
@@ -736,6 +736,17 @@ class SourceColumn(object):
 
         self._calculate_label()
         self._calculate_help_text()
+        if self.is_sortable:
+            field_name = self.sort_field or self.attribute
+            try:
+                get_related_field(
+                    model=self.source, related_field_name=field_name
+                )
+            except FieldDoesNotExist as exception:
+                raise ImproperlyConfigured(
+                    '"{}" is not a field or "{}", cannot be used as a '
+                    'sortable column.'.format(field_name, self.source)
+                ) from exception
 
     def _calculate_help_text(self):
         if not self._help_text: